Remove adaptation options UI, assets page
