--- src/config/domains/adaptation-sector-layers.json
@@ -1,344 +1,0 @@
-[
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Commercial",
-    "layer_name": "buildings_commercial"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Industrial",
-    "layer_name": "buildings_industrial"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Institutional",
-    "layer_name": "buildings_institutional"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Mixed Use",
-    "layer_name": "buildings_mixed"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Other",
-    "layer_name": "buildings_other"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Recreation",
-    "layer_name": "buildings_recreation"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Residential",
-    "layer_name": "buildings_residential"
-  },
-  {
-    "sector": "buildings",
-    "subsector": "buildings",
-    "asset_type": "Resort",
-    "layer_name": "buildings_resort"
-  },
-  {
-    "sector": "power",
-    "subsector": "demand",
-    "asset_type": "demand",
-    "layer_name": "elec_nodes_demand"
-  },
-  {
-    "sector": "power",
-    "subsector": "generation",
-    "asset_type": "diesel",
-    "layer_name": "elec_nodes_diesel"
-  },
-  {
-    "sector": "power",
-    "subsector": "generation",
-    "asset_type": "gas",
-    "layer_name": "elec_nodes_gas"
-  },
-  {
-    "sector": "power",
-    "subsector": "generation",
-    "asset_type": "hydro",
-    "layer_name": "elec_nodes_hydro"
-  },
-  {
-    "sector": "power",
-    "subsector": "generation",
-    "asset_type": "solar",
-    "layer_name": "elec_nodes_solar"
-  },
-  {
-    "sector": "power",
-    "subsector": "generation",
-    "asset_type": "wind",
-    "layer_name": "elec_nodes_wind"
-  },
-  {
-    "sector": "power",
-    "subsector": "transmission",
-    "asset_type": "High Voltage",
-    "layer_name": "elec_edges_high"
-  },
-  {
-    "sector": "power",
-    "subsector": "transmission",
-    "asset_type": "Low Voltage",
-    "layer_name": "elec_edges_low"
-  },
-  {
-    "sector": "power",
-    "subsector": "transmission",
-    "asset_type": "pole",
-    "layer_name": "elec_nodes_pole"
-  },
-  {
-    "sector": "power",
-    "subsector": "transmission",
-    "asset_type": "substation",
-    "layer_name": "elec_nodes_substation"
-  },
-  {
-    "sector": "transport",
-    "subsector": "air",
-    "asset_type": "runway",
-    "layer_name": "airport_runways"
-  },
-  {
-    "sector": "transport",
-    "subsector": "air",
-    "asset_type": "terminal",
-    "layer_name": "airport_terminals"
-  },
-  {
-    "sector": "transport",
-    "subsector": "port",
-    "asset_type": "break",
-    "layer_name": "port_areas_break"
-  },
-  {
-    "sector": "transport",
-    "subsector": "port",
-    "asset_type": "container",
-    "layer_name": "port_areas_container"
-  },
-  {
-    "sector": "transport",
-    "subsector": "port",
-    "asset_type": "industry",
-    "layer_name": "port_areas_industry"
-  },
-  {
-    "sector": "transport",
-    "subsector": "port",
-    "asset_type": "silo",
-    "layer_name": "port_areas_silo"
-  },
-  {
-    "sector": "transport",
-    "subsector": "rail",
-    "asset_type": "junction",
-    "layer_name": "rail_nodes"
-  },
-  {
-    "sector": "transport",
-    "subsector": "rail",
-    "asset_type": "station",
-    "layer_name": "rail_stations"
-  },
-  {
-    "sector": "transport",
-    "subsector": "rail",
-    "asset_type": "track",
-    "layer_name": "rail_edges"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "bridge",
-    "layer_name": "road_bridges"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "CLASS A",
-    "layer_name": "road_edges_class_a"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "CLASS B",
-    "layer_name": "road_edges_class_b"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "CLASS C",
-    "layer_name": "road_edges_class_c"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "METRO",
-    "layer_name": "road_edges_metro"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "OTHER",
-    "layer_name": "road_edges_other"
-  },
-  {
-    "sector": "transport",
-    "subsector": "road",
-    "asset_type": "TRACK",
-    "layer_name": "road_edges_track"
-  },
-  {
-    "sector": "water",
-    "subsector": "irrigation",
-    "asset_type": "canal",
-    "layer_name": "water_irrigation_edges"
-  },
-  {
-    "sector": "water",
-    "subsector": "irrigation",
-    "asset_type": "well",
-    "layer_name": "water_irrigation_nodes"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Booster Station",
-    "layer_name": "water_potable_nodes_booster"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Catchment",
-    "layer_name": "water_potable_nodes_catchment"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Entombment",
-    "layer_name": "water_potable_nodes_entombment"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Filter Plant",
-    "layer_name": "water_potable_nodes_filter"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Intake",
-    "layer_name": "water_potable_nodes_intake"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "potable",
-    "layer_name": "water_potable_edges"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Production Well",
-    "layer_name": "water_potable_nodes_well"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Pump Station",
-    "layer_name": "water_potable_nodes_pump"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Relift Station",
-    "layer_name": "water_potable_nodes_relift"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Reservoir",
-    "layer_name": "water_potable_nodes_reservoir"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "River Source",
-    "layer_name": "water_potable_nodes_river_source"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Spring",
-    "layer_name": "water_potable_nodes_spring"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Storage Tank",
-    "layer_name": "water_potable_nodes_tank"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Sump",
-    "layer_name": "water_potable_nodes_sump"
-  },
-  {
-    "sector": "water",
-    "subsector": "potable",
-    "asset_type": "Treatment Plant",
-    "layer_name": "water_potable_nodes_tp"
-  },
-  {
-    "sector": "water",
-    "subsector": "wastewater",
-    "asset_type": "sewer_gravity",
-    "layer_name": "water_waste_sewer_gravity"
-  },
-  {
-    "sector": "water",
-    "subsector": "wastewater",
-    "asset_type": "sewer_pressure",
-    "layer_name": "water_waste_sewer_pressure"
-  },
-  {
-    "sector": "water",
-    "subsector": "wastewater",
-    "asset_type": "WW Pump Station",
-    "layer_name": "water_waste_nodes_pump"
-  },
-  {
-    "sector": "water",
-    "subsector": "wastewater",
-    "asset_type": "WW Relift Station",
-    "layer_name": "water_waste_nodes_relift"
-  },
-  {
-    "sector": "water",
-    "subsector": "wastewater",
-    "asset_type": "WW Sump",
-    "layer_name": "water_waste_nodes_sump"
-  },
-  {
-    "sector": "water",
-    "subsector": "wastewater",
-    "asset_type": "WW Treatment Plant",
-    "layer_name": "water_waste_nodes_wwtp\n"
-  }
-]
