+++ src/config/networks/styles.ts
@@ -9,8 +9,4 @@ export const NETWORK_STYLES = makeConfig([
     id: 'damages',
     label: 'Damages',
   },
-  {
-    id: 'adaptation',
-    label: 'Adaptation Options',
-  },
 ]);
