+++ src/state/data-selection/networks/network-selection.ts
@@ -1,7 +1,11 @@
 import _ from 'lodash';
-import { atom, selector } from 'recoil';
+import { TransactionInterface_UNSTABLE, atom, selector } from 'recoil';
 
-import { CheckboxTreeState, buildTreeConfig } from '@/lib/controls/checkbox-tree/CheckboxTree';
+import {
+  CheckboxTreeState,
+  buildTreeConfig,
+  recalculateCheckboxStates,
+} from '@/lib/controls/checkbox-tree/CheckboxTree';
 
 import { NETWORK_LAYERS_HIERARCHY } from '@/config/networks/hierarchy';
 import { NetworkLayerType } from '@/config/networks/metadata';
@@ -31,3 +35,20 @@ export const networkSelectionState = selector<NetworkLayerType[]>({
     ) as NetworkLayerType[];
   },
 });
+
+export function syncInfrastructureSelectionStateEffect(
+  { get, set }: TransactionInterface_UNSTABLE,
+  layers: string[],
+) {
+  const currentSelection = get(networkTreeCheckboxState);
+  const updatedTreeState = {
+    checked: {
+      ..._.mapValues(currentSelection.checked, () => false),
+      ..._.fromPairs(layers.map((layer) => [layer, true])),
+    },
+    indeterminate: {},
+  };
+  const resolvedTreeState = recalculateCheckboxStates(updatedTreeState, networkTreeConfig);
+
+  set(networkTreeCheckboxState, resolvedTreeState);
+}
