+++ src/state/layers/data-layers/networks.ts
@@ -5,7 +5,6 @@ import { StyleParams, ViewLayer } from '@/lib/data-map/view-layers';
 import { infrastructureViewLayer } from '@/config/networks/infrastructure-view-layer';
 import { sidebarPathVisibilityState } from '@/sidebar/SidebarContent';
 import { damageMapStyleParamsState } from '@/state/data-selection/damage-mapping/damage-style-params';
-import { adaptationStyleParamsState } from '@/state/data-selection/networks/adaptations';
 import { networkSelectionState } from '@/state/data-selection/networks/network-selection';
 import { networksStyleState } from '@/state/data-selection/networks/networks-style';
 
@@ -25,8 +24,6 @@ export const networkStyleParamsState = selector<StyleParams>({
     switch (get(networksStyleState)) {
       case 'damages':
         return get(damageMapStyleParamsState);
-      case 'adaptation':
-        return get(adaptationStyleParamsState);
       default:
         return {};
     }
